added blacklist option
