+++ plugins/nfoSceneParser/reParser.py
@@ -1,0 +1,19 @@
+def find_re_file(scene_path):
+    return
+
+
+def parse(scene_path):
+    # TODO: parse pattern...
+    file_data = {
+        # "file": nfo_file,
+        "source": "re",
+        # "title": file_title,
+        # "details": file_details,
+        # "studio": file_studio,
+        # "movie": file_movie,
+        # "date": file_date,
+        # "actors": file_actors,
+        # "tags": file_tags,
+        # "rating": file_rating
+    }
+    return file_data
